+++ activerecord/lib/active_record/persistence.rb
@@ -1265,8 +1265,7 @@ module ActiveRecord
 
     def _raise_record_not_destroyed
       @_association_destroy_exception ||= nil
-      key = self.class.primary_key
-      raise @_association_destroy_exception || RecordNotDestroyed.new("Failed to destroy #{self.class.model_name.human} with #{key}=#{send(key)}", self)
+      raise @_association_destroy_exception || RecordNotDestroyed.new(nil, self)
     ensure
       @_association_destroy_exception = nil
     end