+++ activerecord/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Allow to override RecordNotDestroyed message
+
+    Now it can be overridden with I18n key
+    `activerecord.errors.messages.record_not_destroyed`.
+
+    *Enmy Pérez*
+
 *   Fix previous change tracking for `ActiveRecord::Store` when using a column with JSON structured database type
 
     Before, the methods to access the changes made during the last save `#saved_change_to_key?`, `#saved_change_to_key`, and `#key_before_last_save` did not work if the store was defined as a `store_accessor` on a column with a JSON structured database type